+++ packages/plugins/IconMenu/src/language.ts
@@ -10,6 +10,7 @@ const language: LanguageOption[] = [
           hints: {
             attributions: 'Quellennachweis',
             draw: 'Zeichenwerkzeuge',
+            measure: 'Entfernung messen',
             filter: 'Filter',
             layerChooser: 'Kartenauswahl',
             gfi: 'Objektliste',
@@ -27,6 +28,7 @@ const language: LanguageOption[] = [
           hints: {
             attributions: 'Attributions',
             draw: 'Draw tools',
+            measure: 'Measure distance',
             filter: 'Filter',
             layerChooser: 'Choose map',
             gfi: 'Feature list',